move result component to shell
