+++ src/app/footer/footer.component.html
@@ -2,7 +2,7 @@
     <a target="_blank" rel="noreferrer" href="https://mon.gov.ua/">
       <img src="../../assets/images/MES.png" alt="Ministry of Education">
     </a>
-    <a target="_blank" rel="noreferrer"  href="https://www.softserveinc.com/">
-      <img src="../../assets/images/SS.png"   alt="SoftServe">
+    <a target="_blank" rel="noreferrer" href="https://www.softserveinc.com/">
+      <img src="../../assets/images/SS.png" alt="SoftServe">
     </a>
 </div>
