+++ src/app/shell/main/main.component.scss
@@ -26,9 +26,6 @@
     color: white;
   }
   &_question{
-    // font-weight: bold;
-    // font-size: 24px;
-    // line-height: 27px;
     color: #FFFFFF;
   }
   &_hint{
