+++ src/app/shell/result/map/map.component.spec.ts
