+++ src/app/shell/result/map/map.component.ts
