+++ src/app/shell/result/ordering/ordering.component.scss
