+++ src/app/shell/result/organization-cards-list/organization-cards-list.component.html
