+++ src/app/shell/result/result.component.scss
