+++ src/app/shell/result/result.component.spec.ts
