+++ src/app/shell/shell-routing.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {MainComponent} from './main/main.component';
-import {ResultComponent} from './main/result/result.component';
+import {ResultComponent} from './result/result.component';
 import {ProviderGuard} from './provider/provider.guard';
 import {ParentGuard} from './parent/parent.guard';
 import {Routes, RouterModule} from '@angular/router';
